Add files via upload
This project was coded in Haskell, a language that gave me a real hatred for functional programming.  Just a head's up, this project doesn't work correctly.  After struggling with this for weeks, I finally gave it to my professor to ask if they could figure out what was wrong with it.  They never figured it out either.  If I remember correctly, the program was supposed to find polypolynomial numbers within a given range.