Add files via upload
This was my second project for the course, and is the reason I fell in love with Fortran.  Fortran is the only language I've seen that provides a convenient and intuitive way of returning multiple values from a function without having to use arrays, tuples, or any other structure.  This program is essentially a sorting algorithm, using the nearest neighbors algorithm.